add some exception

+++ ICalc.cs
@@ -3,10 +3,10 @@
     internal interface ICalc
     {
         double Result { get; set; }
-        void Sum(int x);
-        void Sub(int x);
-        void Multiply(int x);
-        void Divide(int x);
+        void Sum(double x);
+        void Sub(double x);
+        void Multiply(double x);
+        void Divide(double x);
         event EventHandler<EventArgs> MyEventHandler;
         void CancelLast() { }
     }
